refactor invalid value tests
Several tests are aimed at the handling of invalid values, resulting in
duplication of test code for each function that needs to be tested. This
change adds a common location for invalid values, so that these tests can
simply iterate through all invalid items instead of re-implementing each
test.

The helper function load_corpus_folder work generally, and this change uses
it to refactor several tests for invalid param names.